no functional changes; renaming, moving files, a bit of linting
