+++ backend/src/models/user/user.test.js
@@ -1,5 +1,5 @@
 const {User} = require("./user")
-const {createLanguages} = require("../../utils/createLanguages")
+const {createLanguages} = require("../../utils/languages/createLanguages")
 const validators = require("./user.test.validators")
 const {dbConnect, dbDisconnectAndWipe} = require("../../utils/test-utils/db-handler")
 const testUsers = require("./user.test.data")
